+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 New features:
 
-- Added the `workloadIdentityHost` connection option that overrides the STS host used by the AWS Workload Identity flows, so AWS partitions the connector does not know about can be reached without a connector release. The value is used as given and defaults to the regional `sts.<region>.amazonaws.com` (snowflakedb/snowflake-connector-nodejs#1479)
+- Added the `workloadIdentityHost` connection option that overrides the STS host used by the AWS Workload Identity flows, so AWS partitions unknown to the driver can be reached. The value is used as given; when not set, the STS host is derived from the AWS region (snowflakedb/snowflake-connector-nodejs#1479)
 
 Dependencies: